% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autostsm-package.R
\docType{package}
\name{autostsm-package}
\alias{autostsm}
\alias{autostsm-package}
\title{autostsm: Automatic Structural Time Series Models}
\description{
Automatic model selection for structural time series decomposition into trend, cycle, and seasonal components, plus optionality for structural interpolation, using the Kalman filter. Koopman, Siem Jan and Marius Ooms (2012) "Forecasting Economic Time Series Using Unobserved Components Time Series Models" \doi{10.1093/oxfordhb/9780195398649.013.0006}. Kim, Chang-Jin and Charles R. Nelson (1999) "State-Space Models with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications" \doi{10.7551/mitpress/6444.001.0001}\url{http://econ.korea.ac.kr/~cjkim/}.
}
\author{
\strong{Maintainer}: Alex Hubbard \email{hubbard.alex@gmail.com}

}
\keyword{internal}
