% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{autostsm}
\alias{autostsm}
\title{AutoSTSM}
\description{
\emph{autostsm} Automatic model selection for structural time series decomposition into 
trend, cycle, and seasonal components using the Kalman filter.
See the package vignette using \code{vignette("autostsm")} or \code{browseVignettes("autostsm")} 
to view it in your browser.
}
\author{
Alex Hubbard
}
