% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DGS5}
\alias{DGS5}
\title{5 Year Treasury Yield}
\format{
data.table with columns DATE and DGS5, monthly frequency
}
\source{
FRED
}
\usage{
data(DGS5)
}
\description{
5 Year Treasury Yield
}
\keyword{datasets}
