% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{US GDP Seasonally Adjusted}
\format{
data.table with columns DATE and GDP, quarterly frequency
}
\source{
FRED
}
\usage{
data(GDP)
}
\description{
US GDP Seasonally Adjusted
}
\keyword{datasets}
