% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NA000334Q}
\alias{NA000334Q}
\title{US GDP Not Seasonally Adjusted}
\format{
data.table with columns DATE and NA000334Q, quarterly frequency
}
\source{
FRED
}
\usage{
data(NA000334Q)
}
\description{
US GDP Not Seasonally Adjusted
}
\keyword{datasets}
