% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_ssm.R
\name{stsm_bdiag}
\alias{stsm_bdiag}
\title{Build a block diagonal matrix from two matrices}
\usage{
stsm_bdiag(A, B)
}
\arguments{
\item{A}{The top left matrix}

\item{B}{The bottom right matrix}
}
\value{
A block diagonal matrix
}
\description{
Build a block diagonal matrix from two matrices
}
