% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_build_dates.R
\name{stsm_build_dates}
\alias{stsm_build_dates}
\title{Build the date sequence as a Date type}
\usage{
stsm_build_dates(y)
}
\arguments{
\item{y}{a list object created from stsm_detect_frequency}
}
\value{
a list with the univariate time series and corrected dates
}
\description{
Build the date sequence as a Date type
}
