% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_data_checks.R
\name{stsm_check_exo}
\alias{stsm_check_exo}
\title{Data check for input exo}
\usage{
stsm_check_exo(exo, y)
}
\arguments{
\item{exo}{matrix of exogenous data}

\item{y}{input data y}
}
\value{
none
}
\description{
Checks for proper input of the table exo
}
