% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_data_checks.R
\name{stsm_check_exo_fc}
\alias{stsm_check_exo_fc}
\title{Data check for input exo.fc}
\usage{
stsm_check_exo_fc(exo.fc, n.ahead)
}
\arguments{
\item{exo.fc}{exogenous forecast data}

\item{n.ahead}{forecast periods}
}
\value{
none
}
\description{
Checks for proper input of the table exo.fc
}
