% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_data_checks.R
\name{stsm_check_y}
\alias{stsm_check_y}
\title{Data check for input y}
\usage{
stsm_check_y(y)
}
\arguments{
\item{y}{input data y}
}
\value{
none
}
\description{
Checks for proper input of the table y
}
