% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_detect_seasonality.R
\name{stsm_detect_seasonality}
\alias{stsm_detect_seasonality}
\title{Detect seasonality from the data}
\usage{
stsm_detect_seasonality(
  y,
  freq,
  sig_level = 1e-04,
  prior = NULL,
  full = F,
  spectrum = NULL
)
}
\arguments{
\item{y}{Univariate time series of data values.}

\item{freq}{Frequency of the data (1 (yearly), 4 (quarterly), 12 (monthly), 365.25/7 (weekly), 365.25 (daily))}

\item{sig_level}{Significance level to determine statistically significant seasonal frequencies}

\item{prior}{A data table created from stsm_prior}

\item{full}{Whether to check for full spectrum of seasonality rather than weekly and yearly suggestion}

\item{spectrum, }{Vector of seasonal periods to check}
}
\value{
Numeric vector of seasonal periodicities
}
\description{
Detect seasonality from the data
}
\examples{
\dontrun{
#GDP Not seasonally adjusted
library(autostsm)
data("NA000334Q", package = "autostsm") #From FRED
NA000334Q = data.table(NA000334Q, keep.rownames = TRUE)
colnames(NA000334Q) = c("date", "y")
NA000334Q[, "date" := as.Date(date)]
NA000334Q[, "y" := as.numeric(y)]
NA000334Q = NA000334Q[date >= "1990-01-01", ]
seasonality = stsm_detect_seasonality(y = NA000334Q$y, freq = 4)
}
}
