% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_harmonics.R
\name{stsm_harmonics}
\alias{stsm_harmonics}
\title{Get harmonics for the seasons}
\usage{
stsm_harmonics(freq, decomp = "", seasons = NULL, wavelet = FALSE)
}
\arguments{
\item{freq}{Frequency of the data}

\item{decomp}{Decomposition model ("tend-cycle-seasonal", "trend-seasonal", "trend-cycle", "trend-noise")}

\item{seasons}{The seasonal periods: i.e. c(365.25, 7 if yearly and weekly seasonality). Default is NULL and will be estimated via wavelet analysis.}

\item{wavelet}{Whether seasons were detected using wavelet analysis}
}
\value{
list containing the initial values for the Kalman filter
}
\description{
Get harmonics for the seasons
}
