% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_na_kalman.R
\name{stsm_na_kalman}
\alias{stsm_na_kalman}
\title{Missing Value Imputation by Kalman Smoothing and State Space Models}
\usage{
stsm_na_kalman(y)
}
\arguments{
\item{y}{Univariate time series}
}
\description{
Simplified version taken from the `imputeTS` package.
Uses Kalman Smoothing on structural time series models for imputation. It uses "StructTS" to build a
"basic structural model" if the frequency of y is greater than 1. Otherwise, it uses a local trend model.
}
