% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_dcf}
\alias{sw_dcf}
\title{Stock & Watson US Dynamic Common Factor (1991) Data}
\format{
data.table with columns date, ip (industrial production), 
gmyxpg (personal income less transfer payments), 
mtq (total manufacturing and trade sales), and 
lpnag (employeess on nonagricultural payrolls)
taken from
}
\source{
Kim and Nelson (1999) "State-Space Modesl with Regime Switching" (http://econ.korea.ac.kr/~cjkim/)
}
\usage{
data(sw_dcf)
}
\description{
Stock & Watson US Dynamic Common Factor (1991) Data
}
\keyword{datasets}
