% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNRATE}
\alias{UNRATE}
\title{Unemployment Rate Seasonally Adjusted}
\format{
data.table with columns DATE and UNRATE, monthly frequency
}
\source{
FRED
}
\usage{
data(UNRATE)
}
\description{
Unemployment Rate Seasonally Adjusted
}
\keyword{datasets}
