% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNRATENSA}
\alias{UNRATENSA}
\title{Unemployment Rate Not Seasonally Adjusted}
\format{
data.table with columns DATE and UNRATENSA, monthly frequency
}
\source{
FRED
}
\usage{
data(UNRATENSA)
}
\description{
Unemployment Rate Not Seasonally Adjusted
}
\keyword{datasets}
