% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_data_checks.R
\name{stsm_format_exo}
\alias{stsm_format_exo}
\title{Format exo}
\usage{
stsm_format_exo(exo_obs, exo_state, dates, range)
}
\arguments{
\item{exo_obs}{exogenous observation data}

\item{exo_state}{exogenous state data}

\item{dates}{dates vector}

\item{range}{range of data to include}
}
\value{
a data table
}
\description{
Format the exo table
}
