% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm_init_pars.R
\name{stsm_init_pars}
\alias{stsm_init_pars}
\title{Get initial parameter estimates for estimation}
\usage{
stsm_init_pars(
  y,
  freq,
  trend,
  cycle,
  decomp = "",
  harmonics = NULL,
  prior = NULL
)
}
\arguments{
\item{y}{an object created from stsm_detect_frequency}

\item{freq}{Frequency of the data}

\item{trend}{Trend specification ("random-walk", "random-walk-drift", "double-random-walk", "random-walk2").}

\item{cycle}{The period for the longer-term cycle}

\item{decomp}{Decomposition model ("tend-cycle-seasonal", "trend-seasonal", "trend-cycle", "trend-noise")}

\item{harmonics}{The harmonics to split the seasonality into}

\item{prior}{A data table created by stsm_prior}
}
\value{
named vector containing the initial parameter estimates for estimation
}
\description{
Get initial parameter estimates for estimation
}
